Move list_secret integration test into test_secrets
